core/requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
APScheduler==3.11.0
asyncpg==0.30.0
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
frozenlist==1.5.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
gspread==6.1.4
httplib2==0.22.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
oauth2client==4.1.3
oauthlib==3.2.2
propcache==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dotenv==1.0.1
redis==7.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
yarl==1.18.3
